Chore refactor features
Split features parsing to multiple modules and make pop features usage
consistent across the `utoipa-gen` crate.
